[mc] replace timer.hpp, and remove expl constr
