+++ triqs/utility/timer.hpp
@@ -3,7 +3,7 @@
  *
  * TRIQS: a Toolbox for Research in Interacting Quantum Systems
  *
- * Copyright (C) 2011 by M. Ferrero, O. Parcollet
+ * Copyright (C) 2017 by Hugo U.R. Strand
  *
  * TRIQS is free software: you can redistribute it and/or modify it under the
  * terms of the GNU General Public License as published by the Free Software
@@ -19,28 +19,37 @@
  * TRIQS. If not, see <http://www.gnu.org/licenses/>.
  *
  ******************************************************************************/
+#pragma once
 
-#ifndef TIMER_OP_H_238rj238rh
-#define TIMER_OP_H_238rj238rh  
+#include <chrono>
 
-#include <time.h>
+namespace triqs {
+  namespace utility {
 
-namespace triqs { 
-namespace utility { 
+    class timer {
+      public:
+      using clock_t = std::chrono::high_resolution_clock;
 
-class timer {
-  clock_t Clock1,Clock2;
-  bool running;
-public:
-  timer():running(false) {}
-  void start() { running= true; Clock1 = clock();}
-  void stop() { Clock2 = clock(); running = false;}
-  operator double() const {return  (Clock2 - Clock1)/CLOCKS_PER_SEC;}  
-  };
+      private:
+      clock_t::time_point start_time;
+      clock_t::duration total_time = clock_t::duration(0);
+      bool running = false;
+
+      public:
+      void start() {
+        running    = true;
+        start_time = clock_t::now();
+      }
+      void stop() {
+        total_time += clock_t::now() - start_time;
+        running = false;
+      }
+      bool is_running() const { return running; }
+      operator double() const {
+        std::chrono::duration<double> total_time_seconds(total_time);
+        if (is_running()) total_time_seconds += clock_t::now() - start_time;
+        return total_time_seconds.count();
+      }
+    };
+  }
 }
-}
-
-
-#endif
-
-
